Cambio de nomenclatura en proyectos

+++ server-springboot/src/main/java/com/ccsw/tutorial/category/CategoryServiceImpl.java
@@ -1,18 +1,19 @@
 package com.ccsw.tutorial.category;
 
-import java.util.List;
-
+import com.ccsw.tutorial.category.model.Category;
+import com.ccsw.tutorial.category.model.CategoryDto;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ccsw.tutorial.category.model.Category;
-import com.ccsw.tutorial.category.model.CategoryDto;
+import java.util.List;
 
 /**
-* @author ccsw
-*
-*/
+ * @author ccsw
+ *
+ */
 @Service
+@Transactional
 public class CategoryServiceImpl implements CategoryService {
 
    @Autowired
@@ -28,8 +29,8 @@ public class CategoryServiceImpl implements CategoryService {
    }
 
    /**
-   * {@inheritDoc}
-   */
+    * {@inheritDoc}
+    */
    @Override
    public List<Category> findAll() {
 
@@ -37,30 +38,35 @@ public class CategoryServiceImpl implements CategoryService {
    }
 
    /**
-   * {@inheritDoc}
-   */
+    * {@inheritDoc}
+    */
    @Override
    public void save(Long id, CategoryDto dto) {
 
-      Category categoria = null;
+      Category category;
 
-      if (id == null)
-         categoria = new Category();
-      else
-         categoria = this.get(id);
+      if (id == null) {
+         category = new Category();
+      } else {
+         category = this.get(id);
+      }
 
-      categoria.setName(dto.getName());
+      category.setName(dto.getName());
 
-      this.categoryRepository.save(categoria);
+      this.categoryRepository.save(category);
    }
 
    /**
-   * {@inheritDoc}
-   */
+    * {@inheritDoc}
+    */
    @Override
-   public void delete(Long id) {
+   public void delete(Long id) throws Exception {
+
+      if(this.get(id) == null){
+         throw new Exception("Not exists");
+      }
 
       this.categoryRepository.deleteById(id);
-
    }
+
 }
